ENH: Prevent crash when an invalid outgoing node is added to the connector node
The connector node currently only supports a limited number of outgoing node types (Volumes, Models, Transforms, Text, etc.).
Some nodes, such as vtkMRMLAnnotationRulerNode are technically "vtkMRMLModelNode", and so are listed in the node selector, but caused Slicer to crash when added as an outgoing node.
This commit should fix the crash and log an error message when an invalid outgoing node reference is added.

re #96