ENH: Close running Plus servers when Slicer closes
Before the application closes the modulesAboutToBeUnloaded() is emitted by the qSlicerModuleFactoryManager.
Running PlusServers must be closed at this step, since if the OpenIGTLinkIF module is removed before the PlusRemote module, all of the IGTL connectors will already be deactivated, and we will not be able to communicate with the PlusServerLauncher.